Add util method to sort API resources according to a defined order
