+++ features/admin.api-resources.v2/utils/api-resource-utils.ts
@@ -18,6 +18,7 @@
 
 import { FeatureConfigInterface } from "@wso2is/admin.core.v1/models/config";
 import { hasRequiredScopes } from "@wso2is/core/helpers";
+import { DropdownItemProps } from "semantic-ui-react";
 import { APIResourceCategories, APIResourcesConstants } from "../constants/api-resources-constants";
 import { Property } from "../models/api-resources";
 
@@ -156,5 +157,19 @@ export class APIResourceUtils {
         }
 
         return false;
+    };
+
+    public static sortApiResourceTypes = (firstItem: DropdownItemProps, secondItem: DropdownItemProps) => {
+        const apiResourceSortingOrder: APIResourceCategories[] = [
+            APIResourceCategories.BUSINESS,
+            APIResourceCategories.MCP,
+            APIResourceCategories.TENANT,
+            APIResourceCategories.ORGANIZATION
+        ];
+        const aIndex: number = apiResourceSortingOrder.indexOf(firstItem?.type);
+        const bIndex: number = apiResourceSortingOrder.indexOf(secondItem?.type);
+
+        return (aIndex === -1 ? apiResourceSortingOrder.length : aIndex)
+                - (bIndex === -1 ? apiResourceSortingOrder.length : bIndex);
     };
 }
