+++ features/admin.roles.v2/components/wizard-updated/role-permissions/role-permissions.tsx
@@ -455,11 +455,10 @@ export const RolePermissionsList: FunctionComponent<RolePermissionsListProp> =
                             ?.filter((item: DropdownItemProps) =>
                                 item?.type === APIResourceCategories.TENANT ||
                                 item?.type === APIResourceCategories.ORGANIZATION ||
-                                item?.type === APIResourceCategories.BUSINESS
+                                item?.type === APIResourceCategories.BUSINESS ||
+                                item?.type === APIResourceCategories.MCP
                             ).sort((a: DropdownItemProps, b: DropdownItemProps) =>
-                                APIResourceUtils.resolveApiResourceGroup(a?.type)
-                                    ?.localeCompare(APIResourceUtils
-                                        .resolveApiResourceGroup(b?.type))
+                                APIResourceUtils.sortApiResourceTypes(a, b)
                             )
                         }
                         noOptionsText={ t("common:noResultsFound") }
